Add a simple track replay mode.
This basic track replay mode is implemented via the background thread feature to update the viewport center position to follow the points along the track.
It is started via the track context menu.

ATM there are no other GUI controls, but this can be added later. [start/stop control, speed of replay, etc...]
ATM To stop the replay one has to cancel it via the background job control window.
Only one track replay can be in progress at a time.
[It's probably not a good idea to try having the GPS real time tracking on at the same time - as it would flick between the two positions - but no specific check for this made]