Lazy loading of GPX layers
External GPX layers are only loaded if they are visible.  The saves on
file loading times when external files are large.

If a layer is drawn (made visible) or selected, then it is loaded.  The
disadvantage of this is that summary statistics for parent layers will
not take account of external layers that are not loaded.