Fix lazy loading loop bug
When loading some external layers, the load would trigger a redraw which
would trigger another load and so on.  Fix by setting external_loaded to
TRUE before starting the actual load (this causes further attempts to
load to abort).  Once the read is finished, external_loaded is set to
the success value of the read.